+++ tenant_schemas/fields.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.conf import settings
 
-from .utils import get_tenant_field, get_tenant
+from .utils import get_tenant_field
 
 
 class RLSForeignKey(models.ForeignKey):
@@ -12,6 +12,7 @@ def generate_rls_fk_field():
     """
     This method generate the rls foreign key relation field and it aims to unify this definition in a single point
     """
+    from .models import get_tenant
     return RLSForeignKey(
         settings.TENANT_MODEL,
         to_field=get_tenant_field(),
